Added validator using site's html
