+++ haaretz_parser.py
@@ -2,8 +2,7 @@ import collections
 import hashlib
 from datetime import datetime
 
-import bleach
-
+from base_parser import BaseParser
 from rss_parser import RSSParser
 
 HAARETZ_RSS = "https://www.haaretz.co.il/cmlink/1.1617539"
@@ -17,9 +16,12 @@ class HaaretzParser(RSSParser):
     @staticmethod
     def get_source():
         return "Haaretz"
-        
+
     def should_use_first_item_dedup(self):
         return True
+
+    def _validate_change(self, url: str, new: str):
+        return BaseParser.html_validator(url, new)
 
     def entry_to_dict(self, article):
         article_dict = dict()
@@ -34,17 +36,3 @@ class HaaretzParser(RSSParser):
         article_dict['date_time'] = datetime.now(self.tz)
         return article_dict
 
-    @staticmethod
-    def _strip_html(html: str):
-        """
-        a wrapper for bleach.clean() that strips ALL tags from the input
-        """
-        tags = []
-        attr = {}
-        styles = []
-        strip = True
-        return bleach.clean(html,
-                            tags=tags,
-                            attributes=attr,
-                            styles=styles,
-                            strip=strip)
