+++ israel_hayom_parser.py
@@ -2,8 +2,7 @@ import collections
 import hashlib
 from datetime import datetime
 
-import bleach
-
+from base_parser import BaseParser
 from rss_parser import RSSParser
 
 ISRAEL_HAYOM_RSS = "https://www.israelhayom.co.il/rss.xml"
@@ -21,6 +20,9 @@ class IsraelHayomParser(RSSParser):
     def should_use_first_item_dedup(self):
         return True
 
+    def _validate_change(self, url: str, new: str):
+        return BaseParser.html_validator(url, new)
+
     def entry_to_dict(self, article):
         article_dict = dict()
         article_dict['article_id'] = article.guid
@@ -34,17 +36,3 @@ class IsraelHayomParser(RSSParser):
         article_dict['date_time'] = datetime.now(self.tz)
         return article_dict
 
-    @staticmethod
-    def _strip_html(html: str):
-        """
-        a wrapper for bleach.clean() that strips ALL tags from the input
-        """
-        tags = []
-        attr = {}
-        styles = []
-        strip = True
-        return bleach.clean(html,
-                            tags=tags,
-                            attributes=attr,
-                            styles=styles,
-                            strip=strip)
