+++ rss_parser.py
@@ -12,12 +12,7 @@ class RSSParser(BaseParser):
 
     def parse(self):
         r = feedparser.parse(self.url)
-        import urllib.request
-        with urllib.request.urlopen(self.url) as response:
-            data = response.read()
-            import datetime
-            with open(f"./{self.get_source()}_{str(datetime.datetime.now()).replace(':','')}.xml", "wb") as f:
-                f.write(data)
+
         if r is None:
             logging.warning('Empty response RSS')
             return
